+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - The foreign key from `mail_user_preferences.code` requires the referenced column to have a unique index.
 - Added a priority parameter to the content generator replacers register method to ensure the replacers are applied in the correct order. remp/remp#1450
 - Added AnchorWirelinkReplace content generator replacer to support deeplinking in emails. remp/remp#1450
+- Fixed `RulesTrait` not formatting `[caption]` tags in generator with its designated template but rather with `$imageTemplate`. remp/euobserver#181
 
 ## Archive